src/main/java/com/cloudogu/scm/gotenberg/FileResolver.java
/*
 * Copyright (c) 2020 - present Cloudogu GmbH
 *
 * This program is free software: you can redistribute it and/or modify it under
 * the terms of the GNU Affero General Public License as published by the Free
 * Software Foundation, version 3.
 *
 * This program is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
 * details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see https://www.gnu.org/licenses/.
 */

package com.cloudogu.scm.gotenberg;

import sonia.scm.repository.Repository;
import sonia.scm.repository.api.RepositoryService;
import sonia.scm.repository.api.RepositoryServiceFactory;

import jakarta.inject.Inject;
import java.io.IOException;
import java.io.InputStream;

public class FileResolver {

  private final RepositoryServiceFactory repositoryServiceFactory;

  @Inject
  public FileResolver(RepositoryServiceFactory repositoryServiceFactory) {
    this.repositoryServiceFactory = repositoryServiceFactory;
  }

  public InputStream getContent(Repository repository, RepositoryPath repositoryPath) throws IOException {
    try (RepositoryService service = repositoryServiceFactory.create(repository)) {
      return service.getCatCommand()
        .setRevision(repositoryPath.getRevision())
        .getStream(repositoryPath.getPath());
    }
  }

}
